requirements.txt
Flask>=3.0.0
flask-httpauth>=4.8.0
PyYAML>=6.0
Jinja2>=3.1.0
email-validator>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
IMAPClient>=3.0.0
plyer>=2.1.0
